buildsystem: fix GEN messages in verbose mode
So far, setting silent mode was done in configure.ac, and was on by
default for "release" builds and off for "devel" builds.

However, in "release" builds, running "make V=1" would not display the
commands used to generate tsome files (like the ones generated via moc
or sed), because the MAKEFLAGS would always contain -s and it was not
possible to disable this behaviour unless by calling configure again.

Handle the SILENT_MAKEFLAGS like AM_V_GEN et al. are managed, by
diverting it to the correct value without the need to re-run configure.

At the same time, also provide the verbose versions of AM_V_MOC and
AM_V_GPERF (even if that is technically not needed, it's good for
homogeneity with the rest of the automake-generated code).

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>