libs/parser: support gperf v3.1+
As Freddie wrote:

    New version of gperf (3.1, released in January 2017) breaks the
    build of kconfig-frontends both for the last package you uploaded
    (3.12.0.0) and for the HEAD from the repository.

    The problem is that gperf 3.1 now uses "size_t" as the length
    argument for the functions, while earlier versions use "unsigned
    int".

As suggested by Freddie, add a configure check to detect the type used
in gperf declarations. Unlike the patch from Freddie, we use as much
autoconf code as possible, rather than canned shell commands. This makes
the code about twice bigger, but it is full autoconf stuff, so
presumably more portable.

Also add a patch to be applied by scripts/ksync.sh, so that the change
persists with future update from upstream (as so we catch it when
upstream fixes it by themselves).

Reported-by: Freddie Chopin <freddie.chopin@gmail.com>
Inspired-by: https://gist.github.com/FreddieChopin/b2ec415ee4eacbd4b92499bbf4e934be
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>