Makefile: comment tricky yconf condition
The yconf.c file is generated from the yconf.y yacc source; it is the
main and only source file that is actuall compiled, as it #include-s the
other source files (weird, but that's how it's done upstream).

So yconf.c does not need to be listed as a built source; the other two,
generated from lex and gperf source files, do need to be listed because
they are not part of _SOURCES, but are only in _EXTRA_SOURCES.

However, .c files generated from yacc code are nt considerate as being
intermediate files, and thus are not cleaned automatically.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>