Makefile: ensure utils/ exists before writing to it
Since utils/ was changed to not build recursively it's possible that in
out-of-tree builds the utils/ directory hasn't been created when Make runs the
utils/kconfig-tweak target which tries to write to file inside utils/.

This is the case, for example, when using --disable-dependency-tracking.

To ensure this can work, mkdir the directory.

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit 32a4d43dedfd22654f6b9882571ec9c4e03d3e05)
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>